move lineWrapper back into textStyleBuilder
